refactor(frontend): extract datastudio result chart
